chore: type the new modules, and drop a constant the easing made dead
mypy --strict was never clean on this tree - 20 errors before this
release, all in the platform and glue layers. It is now at 18, and every
module added or touched here is clean: annotated mesh accumulators, real
Mesh return types on the beetle's shape builders, a dataclass instead of a
dict of mixed types in the locomotor fixture, and the concrete Gdk event
classes on the brain window's handlers instead of bare object.

LEDGE_ALIGN_LERP went unused when ledge alignment moved to turn_toward.
The renders and all three suites are byte-for-byte unchanged.

+++ desktopfly/behavior.py
@@ -83,7 +83,7 @@ def lag(rate: float, dt: float) -> float:
     per_frame = min(1.0, rate / k.TUNED_HZ)
     if per_frame >= 1.0:
         return 1.0
-    return 1.0 - (1.0 - per_frame) ** (k.TUNED_HZ * dt)
+    return float(1.0 - (1.0 - per_frame) ** (k.TUNED_HZ * dt))
 
 
 class Fly:
